refactor recommendations object out of recViewController
